Implementation of Multiple Linear Regression model on the following dataset
The prediction was not upto the mark with backward elimination method to predict the F/R, therefore this model doesn't gonna work for this dataset. Hence will look up to some other model in future if those can predict accurately or not by learning the correlation between day of week, month and year accordingly.